Move to autoresponder folder to seperate from other more complex commands
